Added checks to prevent volume and other audio parameters being set to invalid values. Fixes T1213 +
Test Plan: NA

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: joshua.claeys, andrea.ellinger

Differential Revision: https://www.zeroengine.io/D287

+++ ZeroLibraries/AudioEngine/Tags.cpp
@@ -664,9 +664,9 @@ namespace Audio
   }
 
   //************************************************************************************************
-  void TagObject::SetInstanceLimit(const float limit)
+  void TagObject::SetInstanceLimit(const int limit)
   {
-    mInstanceLimit = (unsigned)limit;
+    mInstanceLimit = limit;
 
     if (!mIsThreaded)
     {
